diviner: a black-box optimization framework
Summary:
Diviner is a black-box optimization framework that uses Bigmachine
to distribute a large number of computationally expensive trials
across clusters of machines.

Diviner inherits some of Vizier's [1] core concepts, and adds its
own:

- a parameter describes the type and range of values that can
  parameterize the process under optimization;

- a value is a concrete value for a parameter;

- a metric is a numeric value produced by the process under
  optimization;

- an objective is a metric to maximize or minimize;

- a trial a set of values and the metrics that result from
  the process run under these values;

- an oracle is an algorithm that picks a set of trials to
  perform given a history of trials;

- a black box defines the process under optimization, namely:
  how to run a trial given a set of values; and finally

- a study is a black box, an objective, and an oracle.

Studies (and their associated objects) are configured by a
configuration script written in the Starlark configuration
language [2]. Details of the Starlark builtins provided to create
Diviner configurations are described in package
github.com/grailbio/diviner/script.

As an example, here's a Diviner configuration that tries to
maximize the accuracy of a model, parameterized by the optimizer
used and training batch size:

```
	# model_files is the set of (local) files required to run
	# training and evaluation for the model.
	model_files = ["neural_network.py", "utils.py"]

	# run_model returns a run configuration for a set of
	# parameter values.
	def run_model(pvalues):
	  return run_config(
	    local_files=model_files,
	    script="""
	      python neural_network.py \
	        --batch_size=%(batch_size)d \
	        --optimizer=%(optimizer)s
	    """ % pvalues,
	  )

	# Define a black box that defines a set of parameters for
	# neural_network.py, and provides run configs as above.
	neural_network = black_box(
	  name="neural_network",
	  params={
	    "batch_size": discrete(16, 32, 64, 128),
	    "optimizer": discrete("adam", "sgd"),
	  },
	  run=run_model,
	)

	# Study declares a study that attempts to maximize the "acc"
	# metric returned by the model as defined above.
	study("hparam_study",
	  black_box=neural_network,
	  objective=maximize("acc"),
	)
```

With a configuration in hand, the diviner tool is used to conduct trails,
and examine study results.

Usage:
```
	diviner config.dv list
	diviner config.dv run study [-trials N] [-parallelism P]
	diviner config.dv summarize study
```

`diviner config.dv list` lists the studies provided by the
configuration config.dv.

`diviner config.dv run study [-trials N] [-parallelism P]` conducts
N trials of the named study as configured in config.dv, using a
maximum parallelism of P. If N not provided, then all trails are
run, as long as the set of trials are finite. If P is not defined,
then diviner runs all trials in parallel.

`diviner config.dv summarize study` summarizes the named study; it
outputs a TSV of all known trials, specifying parameter values.
The TSV is ordered by the objective.

[1] https://www.kdd.org/kdd2017/papers/view/google-vizier-a-service-for-black-box-optimization
[2] https://docs.bazel.build/versions/master/skylark/language.html

Reviewers: ysaito, joshnewman, onikolic

Reviewed By: ysaito

Subscribers: joshnewman, kdavydov, ayip, dfilippova, vnicula

Differential Revision: https://phabricator.grailbio.com/D24955

fbshipit-source-id: fbe249e