diviner: shared database; single-file database
Summary:
This change formalizes the notion of a study database in Diviner and
provides two implementations of it:

- The "local" database is a single-file local database backed by
Bolt. This database can be used only by a single process at a time
and replaces the current ad-hoc runs file and logs.

- The "dynamodb" database uses a shared DynamoDB table to store state
and logs. This may safely be shared across multiple users and can
also be used to query live run status, etc. (e.g., a leaderboard).
This should make it much simpler to keep track of everything that is
done, and to perform optimization on an incremental basis. The
dynamodb database uses AWS CloudWatch Logs to store and retrieve log
information. This allows us to monitor runs in realtime and also to
search across all of the logs that are generated.

The change also introduces a new subcommand, "diviner logs", to
print logs for a particular run.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: ayip, dfilippova, vnicula, onikolic

Differential Revision: https://phabricator.grailbio.com/D25638

fbshipit-source-id: dce88fa