diviner: allow multiple studies to be performed concurrently
Summary:
This change allows diviner to perform multiple studies concurrently.
The argument to 'run' becomes a regular expression matching the
studies to be performed. Runner.Do is refactored into a shared
management loop; trials are managed separately by the the new methods
Runner.Run and Runner.Round. (This also opens up the possibility of
more flexible run management; not tied to the specifics of the run
loop.)

This arrangement also means that systems are shared: e.g., two
studies making use of the same systems share that system's pools and
limits. This is the correct behavior.

The status pages now reports on all ongoing studies.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: onikolic, dfilippova, vnicula

Differential Revision: https://phabricator.grailbio.com/D25915

fbshipit-source-id: e789ff9