diviner: configuration-free introspection, leaderboards, etc.
Summary:
This change removes the requirement that diviner is always driven by
a shared configuration. Instead, we read all state from the database.
Diving configs are only required when launching new runs.

This change also reworks diviner's command line tooling to be more
flexible and useful for the purposes tracking and maintaining studies.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: ayip, dfilippova, vnicula, onikolic

Differential Revision: https://phabricator.grailbio.com/D26445

fbshipit-source-id: e7c1e70