diviner: indexed queries
Summary:
This change introduces a secondary index to the dynamoDB database
which allows us to efficiently look up entries by time of update.
This in turn allows for instant "pending" queries, and also a new
query option exposed through "diviner list -since", e.g., to get the
set of studies that have been active in the last 6 hours:

	diviner list -since=6h

Or to get all failed runs in the last hour:

	diviner list -runs -state=failure -since=1h

etc.

However the main benefit of this change is that common queries
like "diviner list -runs -state=pending -s" are now near instant.

Reviewers: ysaito

Reviewed By: ysaito

Subscribers: ayip, vnicula, onikolic, dfilippova, joshnewman

Differential Revision: https://phabricator.grailbio.com/D27155

fbshipit-source-id: 7088be4