diviner: optionally list run values in "diviner list"
Summary:
It's often useful to see a run's values when listing them, for
example to look for a run with a specific set of hyperparameters. We
fix this by adding a new flag to "diviner list". We also interpret
values (and metrics) as lists of anchored regular expressions. This
is useful to express multiple groupings of values (or metrics); for
example:

	$ diviner list -runs -values fragment_packing_policy,sample.* ...

Reviewers: joshnewman, ysaito

Reviewed By: joshnewman

Subscribers: ayip, vnicula, onikolic, dfilippova

Differential Revision: https://phabricator.grailbio.com/D28112

fbshipit-source-id: 150339a