diviner: reintroduce run_id arg
Summary:
This changes how run IDs are allocated. They are now first reserved
via a new database method (NextSeq), and then, once a run has been
constructed (where we can pass in the reserved identifier), we populate
the database entry for that run.

This allows us to pass a run id when we construct the run. Propagate
this all the way to the script.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: vnicula, onikolic, dfilippova

Differential Revision: https://phabricator.grailbio.com/D28969

fbshipit-source-id: 3a78d0a