diviner: implement a watch dog
Summary:
This change introduces support for keepalives in diviner. A process
enter a keepalive monitoring regime by issuing the directive (on
standard IO):

	DIVINER: keepalive=duration

If a new keepalive directive is not issued before the provided
duration expires, the process is killed. It is retried up to 5 times
in such a manner. Duration is a string parseable by
time.ParseDuration. If a process is killed, we also discard
the machine that it ran on.

A monitored process can exit the keepalive regime by issuing a
keepalive of duration 0.

This allows for processes that may hang (e.g., Tensorflow+GPU
interaction bugs) to proceed from a previous checkpoint.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: vnicula, onikolic, dfilippova

Differential Revision: https://phabricator.grailbio.com/D29245

fbshipit-source-id: ffbe489