diviner: keep track of restarts; proper re-sync log streams
Summary:
This repairs a few issues that come up with restarts. First, we
weren't re-syncing log tokens properly from cloudwatch; this caused
retried runs to fail to log entirely. Second, we now display the
number of restarts in "diviner info", and also inject a log message
delimiting the tries of a run.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito

Subscribers: vnicula, onikolic, dfilippova

Differential Revision: https://phabricator.grailbio.com/D29510

fbshipit-source-id: 03564c6