diviner: first-class support for replicates
Summary:
This change adds first-class support for replicates to Diviner. If a
study is configured with replicates, each value combination is run
that many times; their metrics are averaged. The study's run function
is given the replicate number, so that this can be used, e.g., to
select outer folds.

Trials are constructed from all available replicates, and a trial is not
considered complete until all of its replicates are computed.
The leaderboard is computed using aggregated trials.

Reviewers: ysaito, dfilippova

Reviewed By: ysaito

Subscribers: ayip, vnicula, onikolic, joshnewman

Differential Revision: https://phabricator.grailbio.com/D29971

fbshipit-source-id: 41350eb