diviner: automatic failed task restart
Summary:
After this change, Diviner automatically restarts failed tasks. When the
oracle requests a set of values, we first inspect our set of failed tasks
to see if there is a (failed) run we can restore. If there is, we do.

In streaming mode, we retry runs that were failed at the beginning
of the stream, but do not (yet) retry runs that fail during the stream.

This helps with applications where each trial is very expensive and
may be usefully resumed (e.g., because checkpoints are keyed on the
run id).

Reviewers: ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D33265

fbshipit-source-id: 13c50b9