diviner: allow resuming a run that uses a map key
Summary:
The problem is that the codebase conflated *Dict and Dict.  Dict.Equal() returns
false if it's passed a *Dict, and that made it impossible for diviner to match a
failed run that used a map hyperparameter value.

Bonus: add "-C dir" flag that causes diviner to chdir to the given dir on start. It makes it much easier to run diviner under a debugger.

Reviewers: marius, sbagaria

Reviewed By: sbagaria

Differential Revision: https://phabricator.grailbio.com/D33880

fbshipit-source-id: 13492c8